Add support for TCP connections
This adds support for reading P1 data from a TCP port. There are devices
which connect to a P1 capable meter, which then present the data via
the network.

Connecting to a TCP source requires specifying the --host and --port
parameters, for the host/IP address and port to connect to.