Add support for mqtt-rate parameter
This add support for the mqtt-rate parameter. Setting this allows the
rate at which messages are sent to MQTT to be decoupled from the rate at
which telegrams are generated by the P1 device.

The default is to send all telegrams generated by the P1 device.

When `mqtt-rate` is set to a value x > 0, at most one message will be
sent to MQTT every x seconds. Telegrams that arrive from the P1 device
in the intermediate period are discarded.